Делит код на модули
